requirements.txt
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
click==8.1.3
decorator==5.1.1
dill==0.3.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipdb==0.13.11
ipython==8.11.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.16.2
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
wrapt==1.15.0
